the start of the new timeline - renamed `Timeline` SystemParam to `TimelineContext` - added trait `Timeline` and structs `NoteTimeline` & `EventTimeline` - removed unsafe code in timeline_tab - moved timeline related stuff to dedicated root module `timeline`
