+++ phichain-editor/src/ui/widgets/mod.rs
@@ -1,3 +1,2 @@
 pub mod beat_value;
 pub mod easing_value;
-pub mod event;
